GENIO: vcodec: mt8395: venc: Fix using a invalid value for type '_Bool'
vcu->abort and is_codec_suspending is bool type, but set value with int type

CR-Id: AUTO00252384

+++ ubuntu/mtk-vcodec-driver/mt8395/venc/venc_common_if.c
@@ -421,7 +421,7 @@ static int venc_get_param(unsigned long handle,
 	if (inst == NULL)
 		return -EINVAL;
 
-	mtk_vcodec_debug(inst, "%s: %d", __func__, type);
+	mtk_vcodec_debug(inst, "%s: %d", __func__, (int)type);
 	inst->vcu_inst.ctx = inst->ctx;
 
 	switch (type) {
